ref: remove allow(dead_code)
